Introduce view context
Generally, the view context has two functions:
(1) It combine access to the view state and the app model.
(2) It contains (usually, user-initiated) actions updating the view and model state.

In this demo app, the view context takes care of the undo manager and registers undo actions for model updates.

Contributes to #2